jsutil: put all js utils in one file
